requirements.txt
playwright
openai
requests
python-dotenv
